remove crash on upgrade due to unexisting file

+++ data/templates/ubuntu-application/upgrade.py
@@ -46,15 +46,16 @@ python_name = templatetools.python_name(project_name)
 ##### 0.4 update
 if project_version < '0.4':
     ## new licensing format
-    bzr_instance = subprocess.Popen(["bzr", "mv", "LICENSE", "COPYING"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    # if file not versionned, try traditional move, (bzr returncode is None if dir not writable)
-    if bzr_instance.returncode == 3 or bzr_instance.returncode is None:
-        try:
-            os.rename('LICENSE', 'COPYING')
-        except OSError, e:
-            if e.errno == 13:
-                sys.stderr.write(_("Can't rename LICENSE file, check your file permission\n"))
-                sys.exit(1)
+    if os.path.isfile("LICENSE"):
+        bzr_instance = subprocess.Popen(["bzr", "mv", "LICENSE", "COPYING"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        # if file not versionned, try traditional move, (bzr returncode is None if dir not writable)
+        if bzr_instance.returncode == 3 or bzr_instance.returncode is None:
+            try:
+                os.rename('LICENSE', 'COPYING')
+            except OSError, e:
+                if e.errno == 13:
+                    sys.stderr.write(_("Can't rename LICENSE file, check your file permission\n"))
+                    sys.exit(1)
     # transition Copyright -> AUTHORS
     if os.path.isfile('AUTHORS'):
         source_file = 'AUTHORS'
